requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
yfinance
joblib
uvicorn
Flask
flasgger
Flask-HTTPAuth
fastapi
python-dotenv
